Refactor visible energy calculation back into SBNOsc. To do this, change core/Event so that reco interactions are differnt from truth interactions (i.e. new RecoInteraction class) and add in userdata fields to RecoInteraction. Updated SBNOsc and ExampleSelection selections to new core format. Put energy calculations back into SBNOsc/ and don't build util/ anymore
