+++ sbnanalysis/CMakeLists.txt
@@ -17,7 +17,7 @@ include_directories(${PROJECT_SOURCE_DIR})
 
 ADD_SUBDIRECTORY(ana)
 ADD_SUBDIRECTORY(core)
-ADD_SUBDIRECTORY(util)
+#ADD_SUBDIRECTORY(util)
 
 file(WRITE ${CMAKE_INSTALL_PREFIX}/bin/setup_sbnanalysis.sh
 "export SBN_ANALYSIS_DIR=${CMAKE_INSTALL_PREFIX}
