+++ sbnanalysis/ana/ExampleAnalysis/ExampleSelection.cxx
@@ -46,7 +46,7 @@ void ExampleSelection::Finalize() {
 }
 
 
-bool ExampleSelection::ProcessEvent(const gallery::Event& ev, std::vector<Event::Interaction>& reco) {
+bool ExampleSelection::ProcessEvent(const gallery::Event& ev, std::vector<Event::RecoInteraction>& reco) {
   if (fEventCounter % 10 == 0) {
     std::cout << "ExampleSelection: Processing event " << fEventCounter << std::endl;
   }
