+++ sbnanalysis/ana/ExampleAnalysis/ExampleSelection.h
@@ -60,7 +60,7 @@ public:
    * \param reco Reconstructed interactions
    * \return True to keep event
    */
-  bool ProcessEvent(const gallery::Event& ev, std::vector<Event::Interaction>& reco);
+  bool ProcessEvent(const gallery::Event& ev, std::vector<Event::RecoInteraction>& reco);
 
 protected:
   unsigned fEventCounter;  //!< Count processed events
