+++ sbnanalysis/ana/SBNOsc/NueSelection.h
@@ -45,7 +45,7 @@ public:
    * \param Reconstructed interactions
    * \return True to keep event
    */
-  bool ProcessEvent(const gallery::Event& ev, std::vector<Event::Interaction>& reco);
+  bool ProcessEvent(const gallery::Event& ev, std::vector<Event::RecoInteraction>& reco);
 
 protected:
   unsigned fEventCounter;  //!< Count processed events
