+++ sbnanalysis/core/ProcessorBase.hh
@@ -14,7 +14,6 @@
 #include "gallery/Event.h"
 #include "Loader.hh"
 #include "Event.hh"
-#include "util/Interaction.hh"
 
 class TBranch;
 class TFile;
@@ -76,10 +75,10 @@ public:
    * \param reco Reco interactions, to be populated by the user
    * \returns True if event passes filter
    */
-  virtual bool ProcessEvent(const gallery::Event& ev, std::vector<Event::Interaction>& reco) = 0;
+  virtual bool ProcessEvent(const gallery::Event& ev, std::vector<Event::RecoInteraction>& reco) = 0;
 
   /** Pointer to reco event information */
-  std::vector<Event::Interaction>* fReco;  //!< Reco interaction list
+  std::vector<Event::RecoInteraction>* fReco;  //!< Reco interaction list
 
 protected:
   /**
