+++ sbnanalysis/util/CMakeLists.txt
@@ -25,17 +25,12 @@ link_directories(
 include_directories(${CMAKE_SOURCE_DIR})
 include_directories(${CMAKE_CURRENT_SOURCE_DIR})
 
-add_library(sbnanalysis_util_Interaction Interaction.cxx)
-target_link_libraries(sbnanalysis_util_Interaction ${ROOT_LIBRARIES})
+add_library(sbnanalysis_util_Covariance Covariance.cxx)
+target_link_libraries(sbnanalysis_util_Covariance ${ROOT_LIBRARIES})
 
-install(TARGETS sbnanalysis_util_Interaction DESTINATION lib)
+add_executable(sbn-cov sbn_cov.cxx)
+target_link_libraries(sbn-cov sbnanalysis_util_Covariance sbnanalysis_Event)
 
-#add_library(sbnanalysis_util_Covariance Covariance.cxx)
-#target_link_libraries(sbnanalysis_util_Covariance ${ROOT_LIBRARIES})
-
-#add_executable(sbn-cov sbn_cov.cxx)
-#target_link_libraries(sbn-cov sbnanalysis_util_Covariance sbnanalysis_Event)
-
-#install(TARGETS sbnanalysis_util_Covariance DESTINATION lib)
-#install(TARGETS sbn-cov DESTINATION bin)
+install(TARGETS sbnanalysis_util_Covariance DESTINATION lib)
+install(TARGETS sbn-cov DESTINATION bin)
